fix(transfer): resolve type mismatches and add build verification guide
- Fix i32 to i16 type mismatch in FundingAdapter (status column)
- Fix i32 to i16 type mismatch in funding/service.rs (account_type column)
- Add binary freshness check to test_transfer_e2e.sh (touch + rebuild if needed)
- Add port conflict resolution with SIGKILL fallback
- Create agent-build-verification-guide.md documenting build and test best practices
- Update SUMMARY.md with new guide link

All E2E and integration tests passing.

+++ src/funding/service.rs
@@ -159,7 +159,7 @@ impl TransferService {
         let mut balances = Vec::new();
         for row in rows {
             let asset_id: i32 = row.get("asset_id");
-            let account_type: i32 = row.get("account_type");
+            let account_type: i16 = row.get("account_type");
             let available: Decimal = row.get("available");
             let frozen: Decimal = row.get("frozen");
             let asset_name: String = row.get("asset_name");
